+++ command/crypto/jwk/create.go
@@ -6,6 +6,7 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
+	"strconv"
 
 	"github.com/pkg/errors"
 	"github.com/smallstep/cli/flags"
@@ -15,6 +16,7 @@ import (
 	"go.step.sm/cli-utils/errs"
 	"go.step.sm/cli-utils/ui"
 	"go.step.sm/crypto/jose"
+	"go.step.sm/crypto/keyutil"
 	"go.step.sm/crypto/randutil"
 )
 
@@ -432,8 +434,9 @@ func createAction(ctx *cli.Context) (err error) {
 		}
 		// If size is not set it will use a safe default
 		if ctx.IsSet("size") {
-			if size < 2048 && !ctx.Bool("insecure") {
-				return errs.MinSizeInsecureFlag(ctx, "size", "2048")
+			minimalSize := keyutil.MinRSAKeyBytes * 8
+			if size < minimalSize && !ctx.Bool("insecure") {
+				return errs.MinSizeInsecureFlag(ctx, "size", strconv.Itoa(minimalSize))
 			}
 			if size <= 0 {
 				return errs.MinSizeFlag(ctx, "size", "0")